refactor: https://github.com/next-step/java-ladder/pull/553#discussion_r476916756 수정
